Move the Wasm module creation code to a new `module` module
