+++ wasm-node/javascript/src/index-deno.ts
@@ -17,7 +17,7 @@
 
 import { Client, ClientOptions, start as innerStart } from './client.js'
 import { Connection, ConnectionConfig } from './instance/instance.js';
-import { default as wasmBase64 } from './instance/autogen/wasm.js';
+import { compileModule } from './module-deno.js'
 
 export {
     AddChainError,
@@ -43,13 +43,7 @@ export {
 export function start(options?: ClientOptions): Client {
     options = options || {};
 
-    // The actual Wasm bytecode is base64-decoded then deflate-decoded from a constant found in a
-    // different file.
-    // This is suboptimal compared to using `instantiateStreaming`, but it is the most
-    // cross-platform cross-bundler approach.
-    const wasmModule = zlibInflate(trustedBase64Decode(wasmBase64)).then(((bytecode) => WebAssembly.compile(bytecode)));
-
-    return innerStart<ParsedAddress>(options || {}, wasmModule, {
+    return innerStart<ParsedAddress>(options || {}, compileModule(), {
         performanceNow: () => {
             return performance.now()
         },
@@ -98,52 +92,6 @@ export function start(options?: ClientOptions): Client {
 type ParsedAddress =
     { ty: "tcp", hostname: string, port: number } |
     { ty: "websocket", url: string }
-
-/**
- * Applies the zlib inflate algorithm on the buffer.
- */
-async function zlibInflate(buffer: Uint8Array): Promise<Uint8Array> {
-    // This code has been copy-pasted from the official streams draft specification.
-    // At the moment, it is found here: https://wicg.github.io/compression/#example-deflate-compress
-    const ds = new DecompressionStream('deflate');
-    const writer = ds.writable.getWriter();
-    writer.write(buffer);
-    writer.close();
-    const output = [];
-    const reader = ds.readable.getReader();
-    let totalSize = 0;
-    while (true) {
-        const { value, done } = await reader.read();
-        if (done)
-            break;
-        output.push(value);
-        totalSize += value.byteLength;
-    }
-    const concatenated = new Uint8Array(totalSize);
-    let offset = 0;
-    for (const array of output) {
-        concatenated.set(array, offset);
-        offset += array.byteLength;
-    }
-    return concatenated;
-}
-
-/**
- * Decodes a base64 string.
- *
- * The input is assumed to be correct.
- */
-function trustedBase64Decode(base64: string): Uint8Array {
-    // This code is a bit sketchy due to the fact that we decode into a string, but it seems to
-    // work.
-    const binaryString = atob(base64);
-    const size = binaryString.length;
-    const bytes = new Uint8Array(size);
-    for (let i = 0; i < size; i++) {
-        bytes[i] = binaryString.charCodeAt(i);
-    }
-    return bytes;
-}
 
 /**
  * Tries to open a new connection using the given configuration.
@@ -488,32 +436,4 @@ declare namespace Deno {
          */
         setKeepAlive(keepalive?: boolean): void;
     }
-}
-
-// Original can be found here: https://github.com/denoland/deno/blob/main/ext/web/lib.deno_web.d.ts
-/**
- * An API for decompressing a stream of data.
- *
- * @example
- * ```ts
- * const input = await Deno.open("./file.txt.gz");
- * const output = await Deno.create("./file.txt");
- *
- * await input.readable
- *   .pipeThrough(new DecompressionStream("gzip"))
- *   .pipeTo(output.writable);
- * ```
- */
-declare class DecompressionStream {
-    /**
-     * Creates a new `DecompressionStream` object which decompresses a stream of
-     * data.
-     *
-     * Throws a `TypeError` if the format passed to the constructor is not
-     * supported.
-     */
-    constructor(format: string);
-
-    readonly readable: ReadableStream<Uint8Array>;
-    readonly writable: WritableStream<Uint8Array>;
 }
