+++ wasm-node/javascript/src/module/autogen/.gitignore
